+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Next
+- Fix undo in the filter textbox not stopping at selection/cursor changes
+
 ## 1.2.0 - 2026 Aug 31
 - Remove macOS 10.15 support
 - Add exact filtering with double quotes, like `"the one"`